gdb/fortran: Don't access non-existent type fields
When attempting to call a Fortran function for which there is no debug
information we currently trigger undefined behaviour in GDB by
accessing non-existent type fields.

The reason is that in order to prepare the arguments for a call to a
Fortran function we need to know about the type of each argument.  If
the function being called has no debug information then obviously GDB
doesn't know about the argument types and we should either give the
user an error or pick a suitable default.  What we currently do is
just assume the fields exist and access them anyway, which is clearly
wrong.

The reason GDB needs to know the argument type is to tell if the
argument is artificial or not, artificial arguments will be passed by
value while non-artificial arguments will be passed by reference.

When considering what was the "correct" default to assume I struggled
to decide, both assumptions seem to have advantages.  For example, if
we assume all arguments are non-artificial and pass them by reference
then passing the "normal" arguments to a function works how the user
expects, for example:

  (gdb) call function_without_debug (1, 2, 3)

will pass the three integers 1, 2, 3.  This is just how we would
expect this to appear in the Fortran source code, and is how we would
call the same function if it did have debug information.

The problem is that if we need to pass an artificial argument by value
then we just can't do it under this scheme.  So, if we instead assume
all arguments are artificial, to call the same function the user would
now have to write:

  (gdb) call function_without_debug (&1, &2, &3)

which is different to how the function is called when it does have
debug information.

My proposal in this commit is to add a new setting that lets the user
pick between these two choices. So we now have:

  set fortran arguments-are-artificial on|off
  show fortran arguments-are-artificial

The default is off, so GDB assumes arguments are not artificial, and
passes by reference, this is the first case highlighted above.

When this setting is turned on we get the second behaviour and
non-artificial arguments need to have their address taken.

gdb/ChangeLog:

	PR fortran/26155
	* NEWS: Mention new setting.
	* f-lang.c: Add 'gdbcmnd.h' include.
	(evaluate_subexp_f): Call fortran_prepare_argument.
	(fortran_arguments_are_artificial): New static global.
	(show_fortran_arguments_are_artificial): New function.
	(set_fortran_list): New static global.
	(show_fortran_list): New static global.
	(_initialize_f_language): Register new commands.
	(fortran_argument_convert): Update comment, make static.
	(fortran_prepare_argument): New function.
	* f-lang.h (fortran_argument_convert): Delete declaration.

gdb/testsuite/ChangeLog:

	PR fortran/26155
	* gdb.fortran/call-no-debug-func.f90: New file.
	* gdb.fortran/call-no-debug-prog.f90: New file.
	* gdb.fortran/call-no-debug.exp: New file.

gdb/doc/ChangeLog:

	PR fortran/26155
	* gdb.texinfo (Fortran): Document new setting.