gdb: Don't hard code 0 as end marker in GDB's line table
Currently GDB hard-codes the use of 0 as the end of sequence marker in
its line tables.  After this commit GDB now uses a named
constant (linetable_entry::end_marker) as the end of sequence marker.

The motivation for this change was discussion on PR gdb/26243.  This
bug discusses the possibility of supporting DWARF's use of line number
0 to indicate an address in the program with no associated source
code.

If we want to have line number 0 and end-of-sequence marker within
GDB's line tables then it seems clear that the end-of-sequence marker
can't take the value 0.  This change is a first step towards changing
the value of the end-of-sequence marker.

Initially I did change the value of the marker in this commit from 0
to -1, however, it was pointed out during review that this causes
problems in GDB when presented with some DWARF that includes use of
line number 0.

The problem is that currently and DWARF that uses line number 0 will
conflict with the end-of-sequence marker, those addresses marked as
line number 0 will (to GDB) appear as end-of-sequence markers.  It
just so happens that in many cases treating an address (with line
number 0) as an end-of-sequence marker is good enough, for example,
end of sequence markers don't have associated source code.

The minute we change the end-of-sequence marker value to -1 the line
number 0 stops being special and we start to run into problems, for
example GDB will try to access line 0 from source files.

It is my proposal that we merge this commit when introduces the
end_marker constant, and makes use of this throughout GDB, but leave
the value of this constant as 0 so that GDB will continue to treat any
DWARF with line number 0 as end-of-sequence markers.

This commit can then be followed up with a series of work to extend
GDB to correctly handle line number 0 in its own right.  Once we
believe that all, or most cases needed to support line number 0 are
being handled then the value of end_marker can be changed from 0 to -1
in a future commit.

For testing this commit I did change the value of end_marker from 0 to
-1 and ran the regression testsuite which passed with no regressions.
This indicates (to me) that I have correctly updated all (or the
majority of) of the hard coded checks for 0 and replaced these with a
check for end_marker instead.

This commit does mention PR gdb/26243 in its ChangeLog, but does not
fix this issue.  I placed this bug reference just so people can see
the connection between this change and that bug in the future.

gdb/ChangeLog:

	PR gdb/26243
	* buildsym.c (buildsym_compunit::record_line): Add an assert for
	the incoming line number.  Update comments to not mention 0
	specifically.  Update to check for linetable_entry::end_marker
	rather than 0.
	(buildsym_compunit::end_symtab_with_blockvector): Check for
	linetable_entry::end_marker, not 0.
	* disasm.c (line_is_less_than): Likewise.
	(do_mixed_source_and_assembly_deprecated): Likewise.
	(do_mixed_source_and_assembly): Likewise.
	* dwarf2/read.c (dwarf_finish_line): Pass
	linetable_entry::end_marker not 0.
	* mdebugread.c (add_line): Set is_stmt field.
	* python/py-linetable.c (ltpy_get_all_source_lines): Check for
	linetable_entry::end_marker not 0, update comments.
	(ltpy_iternext): Likewise.
	* record-btrace.c (btrace_find_line_range): Likewise.
	* symmisc.c (maintenance_print_one_line_table): Likewise.
	* symtab.c (find_pc_sect_line): Likewise.
	(find_line_symtab): Likewise.
	(skip_prologue_using_sal): Likewise.
	* symtab.h (linetable_entry::end_marker): New const static member
	variable.  Add a static assert for this field.
	* xcoffread.c (arrange_linetable): Check for
	linetable_entry::end_marker not 0.