gdb: Change 'set print asm-demangle' to an auto boolean, default on
This commit relates to bug PR gdb/26270.

When I raised this bug it was because I noticed, that when debugging a
C++ application that the 'disassemble' command started displaying
mangled names in its header line.  This was a change from what I was
used to.

I tested GDB 7.6, and confirmed that for at least the last 6 years
disassemble would display demangled names in the header line when
disassembling C++ code.

On further investigation it also turns out that 'info line' output
also displays demangled names, and the symbols referenced in
disassembly output are also demangled for the same time period.

The reason this changed recently is that a long standing bug was fixed
in GDB where (if I understand correctly) GDB was storing the demangled
names where it should have been storing the mangled names, this made
commands line 'set print demangle' and 'set print asm-demangle' not
work, and gave us the behaviour I was seeing on the older GDBs.  With
this bug fix GDB now does what it "should" have all along.

My problem here is that it feels like, if we've lived with things a
particular way for 6 years, then when we change that behaviour it can
seem confusing - it certainly confused me.

So, one reason that we see things being printed as mangled is that
though 'set print demangle' is on by default, 'set print asm-demangle'
is off by default.

So my initial thought was, lets make 'set print asm-demangle' be on by
default.  Then we've still fixed the bug in GDB (about storing the
mangled vs demangled symbols), but the behaviour of GDB doesn't
actually change (vs 9.x release).  After a discussion on IRC the
suggestion was made to have 'set print asm-demangle' be an auto-bool,
then the default can be auto, which would defer to 'set print
demangle', but the user can still take separate control if they want.

So, that's what this patch does.

Unfortunately there's some problems with this approach.

First, and maybe the smallest problem, boolean variables have the
"feature" that a user can just type 'set print asm-demangle', which is
equivalent to 'set print asm-demangle on'.  However, this doesn't work
for auto-booleans, in that case you must type 'set print asm-demangle
on'.

This could possibly be worked around by overriding handling of the
'set print asm-demangle' command, but I haven't done that here.  So
this currently is a breaking change, though a pretty minor one I
think.

The biggest problem, and the blocker I think, is that under testing I
found that the issues I listed above are, apparently, language
specific.  So for 6+ years GDB _for C++_ has been showing demangled
names when it should have been displaying mangled ones, but for Ada,
GDB has been doing the "right" thing.

If I flip the default for 'asm-demangle' on the grounds that I don't
think we should suddenly change GDB's behaviour then this would only
be for C++, for Ada I would be suddenly changing the default.

As a result I don't see any obvious way we could include this patch
without changing GDB's behaviour for one language or another.

What I did discover in my journey into GDB's demangling, is that these
two options are a mess.  The manual entry for 'set print demangle' and
'set print asm-demangle' are super unclear about which parts of GDB
each are supposed to control.  The comment on the controlling
variable (asm_demangle) in the source code (gdb-demangle.h) suggest
this variable should be overridden by 'demangle' if 'demangle' is set
false, however, in some places 'asm_demangle' is completely
independent, and in other places 'asm_demangle' actually overrides
'demangle' rather than deferring to it.

gdb/ChangeLog:

	PR gdb/26270
	* NEWS: Mention change in default for asm-demangle.
	* cli/cli-cmds.c (disassemble_command): Switch to use
	asm_demangle_p.
	* gdb-demangle.c (asm_demangle): Change type to auto bool, default
	is auto.
	(show_asm_demangle): Update to deal with new asm_demangle type.
	(asm_demangle_p): New function.
	(_initialize_gdb_demangle): Register 'set print asm-demangle' as
	an auto-boolean.
	* gdb-demangle.h (asm_demangle): Delete declaration.
	(asm_demangle_p): Declare.
	* printcmd.c (build_address_symbolic): Call asm

gdb/doc/ChangeLog:

	PR gdb/26270
	* gdb.texinfo (Print Settings): Note change in default for
	asm-demangle.  Rewrite the @item line for this option.

gdb/testsuite/ChangeLog:

	PR gdb/26270
	* gdb.ada/win_fu_syms.exp: Update expected output.
	* gdb.base/default.exp: Update expected output.
	* gdb.cp/disasm-func-name.exp: Update for change in asm-demangle,
	add new tests.