gdb: process startup files and startup command line options
Adds the ability to process commands at a new phase during GDB's
startup.  This phase is earlier than the current initialisation file
processing, before GDB has produced any output.

The number of commands that can be processed at this early stage will
be limited, and it is expected that the only commands that would be
processed at this stage will relate to some of the fundamentals of how
GDB starts up.

This commit adds a mechanism to find these startup files as well as
some corresponding command line flags.  Later commits will add
commands that require this functionality in order to do what the user
might want.

The startup files that GDB will currently check for are
~/.config/gdb/gdbstartup (on Linux like systems) or ~/.gdbstartup if
the former is not found.  GDB will also look for ./.gdbstartup, a
local startup file, with the usual auto-loading restrictions.

The output of 'gdb --help' has been extended to include a list of the
startup files being processed, as well as better handling the case
where no startup files are found, so you might now see:

  At startup, GDB reads the following startup files and executes their commands:
     None found.

The printing of 'None found.' has also been applied to .gdbinit files
too.

There are no tests yet for this functionality, as there are no options
that really make use of it, but later commits will add some tests.

gdb/ChangeLog:

	* NEWS: Mention new startup files and command line options.
	* config.in: Regenerate.
	* configure: Regenerate.
	* configure.ac: Define GDBSTARTUP.
	* main.c (relocate_gdbinit_path_maybe_in_datadir): Rename to...
	(relocate_file_path_maybe_in_datadir): ...this.
	(class gdb_initfile_finder): New class.
	(get_init_files): New uses gdb_initfile_finder.
	(get_startup_files): New function.
	(enum cmdarg_kind): Add CMDARG_STARTUP_FILE and
	CMDARG_STARTUP_COMMAND.
	(captured_main_1): Add support for new command line flags, and for
	processing startup files.
	(print_gdb_help): Include startup files in the output, print 'None
	found' when there are no startup files, or no init files.

gdb/doc/ChangeLog:

	* gdb.texinfo (File Options): Mention new command line options.
	(Startup): Discuss when startup files are processed.
	(Initialization Files): Add description of startup files.
	(Auto-loading): Mention startup files.
	(gdb man): Likewise.