gdb: add mechanism to auto-save startup options
As we add settings to GDB (in the following patches) that effect GDB's
startup behaviour a user might desire a mechanism to automatically
save these options into a file.

The reason that this could be important is that, unlike other,
non-startup options, the startup options will only effect GDB's
startup process.  As a result setting the option within a GDB session
will not change how the user experiences that session, but could only
effect the following sessions.

We could take the position that it is entirely the users
responsibility to add these options to a startup file and maintain
them, however, this commit provides a mechanism to automate that
management.

If a user adds the following lines to ~/.config/gdb/gdbstartup:

  set startup-save-filename ~/.config/gdb/auto-save
  source ~/.config/gdb/auto-save

Then GDB will write all startup options to the file auto-save file any
time one of the options is modified.  The source ensures that the
options are reloaded correctly.

This auto saving feature is currently off by default.

There are currently no options that can be auto-saved in this way, but
later commits will add some.

gdb/ChangeLog:

	* cli/cli-setshow.c: Add 'main.h' and 'cli/cli-style.h' include.
	(startup_save_filename): New static global.
	(show_startup_save_filename): New function.
	(write_startup_functions): New static global.
	(write_startup_file): New function.
	(default_startup_writer_callback): New function.
	(add_startup_writer): New function.
	(add_default_startup_writer): New function.
	(_initialize_cli_setshow): New function.
	* cli/cli-setshow.h (write_startup_file): Declare.
	(write_startup_setting_ftype): New typedef.
	(add_startup_writer): Declare.
	(add_default_startup_writer): Declare.
	* main.c (startup_file_read): New static global.
	(gdb_startup_file_read_p): New function.
	(captured_main_1): Set startup_file_read.
	* main.h (gdb_startup_file_read_p): Declare.

gdb/doc/ChangeLog:

	* gdb.texinfo (Commands): Add link to new section.
	(Automatically saving startup options): New section.