gdb: allow default target description to be held on gdbarch
Target descriptions are currently cached in a couple of places:

Calling get_tdesc_info (current_inferior ()) provides access to a
'struct target_desc_info'.  This structure caches the target
description either fetched from the target, or read from a user
supplied file.  This target description is placed inside the 'struct
gdbarch_info' that is then passed to each architectures
'*_gdbarch_init' method when we need to lookup or create a gdbarch
object.

When a gdbarch is created by calling gdbarch_alloc we pass in the
'struct gdbarch_info' and the target description (pointer) is copied
into the 'struct gdbarch'.

And so, once GDB has initialised a gdbarch the target description can
be found in two places, cached off the current inferior, and cached in
the gdbarch.

Currently it is always the case that, if we get a target description
either from the target, or from a user supplied file, then this target
description will be cached on the inferior and will also be cached on
the gdbarch.  In all other cases, both of these caches will contain
nullptr.

As an aside it's not really clear to me what, if anything, are the
benefits of having this information stored in two places.

Now there is one other place where target descriptions can come from,
that is an architectures *_gdbarch_init method.  Many architectures
follow a model of, when a target description is not supplied (from the
target or user) then a default description is looked up, and the
architecture is configured using this default description.

These default target descriptions are never cached on either the
inferior, or in the gdbarch.

A downside to this is that commands like 'maint print xml-tdesc' and
'maint print c-tdesc' will not work if the target is using a default
target description.

This commit prepares gdb to move to a world where these default target
descriptions can be stored on the gdbarch, but not on the inferior.
This means that the maintenance commands mentioned above will now
work.  I find this useful for debugging gdb.

The main idea behind this commit then is to remove the assumption that
if there is no target description cached on the inferior then there
will be no target description on the gdbarch.

This commit does not actually change any targets to store their
default target description within gdbarch, this will come in a later
commit.

gdb/ChangeLog:

	* target-descriptions.c (target_find_description): Add an assert,
	and remove a different assert.
	(target_clear_description): Remove check of target_desc_fetched.
	(tdesc_use_registers): Add an assert.
	(maint_print_tdesc_lookup_tdesc): New function.
	(maint_print_c_tdesc_cmd): Update to use
	maint_print_tdesc_lookup_tdesc.
	(maint_print_xml_tdesc_cmd): Likewise.