gdb: better handling of 'S' packets
This commit builds on work started in the following two commits:

  commit 24ed6739b699f329c2c45aedee5f8c7d2f54e493
  Date:   Thu Jan 30 14:35:40 2020 +0000

      gdb/remote: Restore support for 'S' stop reply packet

  commit cada5fc921e39a1945c422eea055c8b326d8d353
  Date:   Wed Mar 11 12:30:13 2020 +0000

      gdb: Handle W and X remote packets without giving a warning

This is related to how GDB handles remote targets that send back 'S'
packets.

In the first of the above commits we fixed GDB's ability to handle a
single process, single threaded target that sends back 'S' packets.
Although the 'T' packets would always be preferred to 'S' these days,
there's nothing really wrong with 'S' for this situation.

The second commit above fixed an oversight in the first commit, a
single-process, multi-threaded target can send back a process wide
event, for example the process exited event 'W' without including a
process-id, this also is fine as there is no ambiguity in this case.

In PR gdb/26819 however we start to move towards "better" handling of
more ambiguous cases.  In this bug openocd is used to drive the spike
RISC-V simulator.  In this particular case a multi-core system is
being simulated and presented to GDB as two threads.  GDB then is
seeing a single process, two thread system.  Unfortunately the
target (openocd) is still sending back 'S' packets, these are the
packets that _don't_ include a thread-id.

It is my opinion that this target, in this particular configuration,
is broken.  Even though it is possible, by being very careful with how
GDB is configured to ensure that GDB only ever tries to run one thread
at a time, I feel that any target that presents multiple threads to
GDB should be making use of the 'T' stop packet, combined with sending
a thread-id.

However, with that caveat out of the way, I think this bug report does
reveal a number of ways that GDB could be improved.

Firstly, the main issue reported in the bug was that GDB would exit
with this assertion:

  infrun.c:5690: internal-error: int finish_step_over(execution_control_state*): Assertion `ecs->event_thread->control.trap_expected' failed.

I think it's fair to say that having a target send back 'S' packets
when it should use 'T' is _not_ an excuse for GDB to throw an
assertion.

What's happening is that GDB connects to the 2 core system.  Core 2 is
selected, and a program loaded.  A breakpoint is placed in main and we
continue, this results in this packet exchange:

 Sending packet: $vCont;c#a8...Packet received: T05thread:2;

That's good, all cores ran, and the remote told GDB that we stopped in
thread 2.  Next the user does `stepi` and this results in this packet
exchange:

  Sending packet: $vCont;s:2#24...Packet received: S05

Here GDB is trying to step only thread 2, and the target replies with
an 'S' packet.  Though it feels like sending back 'T05thread:2;' would
be so much simpler here, there's nothing fundamentally wrong ambiguous
about the exchange.

Inside GDB the problem we're running into is within the function
remote_target::process_stop_reply.  When a stop reply doesn't include
a thread-id (or process-id) it is this function that is responsible
for looking one up.  Currently we always just select the first
non-exited thread, so in this case thread 1.

As the step was issued as part of the step over a breakpoint logic,
which was specifically run for thread-2 GDB is expecting the event to
be reported in thread-2, and hence when we try to handle thread-1 we
trigger the above assertion.

My proposal is to improve the logic used in process_stop_reply to make
the thread selection smarter.

My first thought was that each thread has an 'executing' flag, instead
of picking the first non-exited thread, we should pick a non-exited
thread that is currently executing.  The logic being that stop events
shouldn't arrive for threads that are no executing.

The problem with this is the very first initial connection.

When GDB first connects to the remote target it is told about all the
existing threads.  These are all created by GDB in the non-executing
state.  Another part of the connecting logic is to send the remote the
'?' packet, which asks why the target halted.  This sends back a stop
packet which is then processed.  At this point non of the threads are
marked executing so we would end up with no suitably matching threads.

This left me with two rules:

  1. Select the first non-exited thread that is marked executing, or
  2. If no threads match rule 1, select the first non-exited thread
  whether it is executing or not.

This seemed fine, and certainly resolved the issue seen in the
original bug report.  So then I tried to create a test for this using
a multi-threaded test program and `gdbserver --disable-packet=T`.

I wasn't able to get anything that exactly reproduced the original
bug, but I was able to hit similar issues where GDB would try to step
one thread but GDB would handle the step (from the step) in a
different thread.  In some of these cases there was genuine ambiguity
in the reply from the target, however, it still felt like GDB could do
a better job at guessing which thread to select for the stop event.

I wondered if we could make use of the 'continue_thread' and/or the
'general_thread' to help guide the choice of thread.

In the end I settled on these rules for thread selection:

  [ NOTE: For all the following rules, only non-exited threads are
    considered.  ]

  1. If the continue_thread is set to a single specific thread, and
  that thread is executing, then assume this is where the event
  occurred.

  2. If nothing matches rule 1, then if the general_thread is set to a
  single specific thread, and that thread is executing, assume this is
  where the event occurred.

  3. If nothing matches rule 2 then select the first thread that is
  marked as executing.

  4. If nothing matches rule 3 then select the first thread.

This works fine except for one small problem, when GDB is using the
vcont packets we don't need to send 'Hc' packets to the target and so
the 'continue_thread' is never set.

In this commit I add a new record_continue_thread function, this sets
the continue_thread without sending a 'Hc' packet.  This effectively
serves as a cache for which thread did we set running.

The only slight "wart" here is that when GDB steps a thread the
continue_thread is not set to a specific single thread-id, rather it
gets set to either minus_one_ptid or to a specific processes ptid.  In
this case (when a step is requested) I store the ptid of the stepping
thread.

gdb/ChangeLog:

	PR gdb/26819
	* remote.c (remote_target::guess_thread_for_ambiguous_stop_reply):
	New member function.
	(record_continue_thread): New function.
	(remote_target::remote_resume_with_vcont): Call record_continue_thread.
	(remote_target::process_stop_reply): Call
	guess_thread_for_ambiguous_stop_reply.

gdb/testsuite/ChangeLog:

	PR gdb/26819
	* gdb.server/stop-reply-no-thread-multi.c: New file.
	* gdb.server/stop-reply-no-thread-multi.exp: New file.