[WIP] gdb: better timeout handling for remote packets
When we get a timeout trying to receive a packet from a remote packet
we currently write 'timeout' into the packet, the result of this is
that the packet is assumed to be OK.

This commit changes it so that the packet is instead treated as
PACKET_UNKNOWN in packet_check_result.  The result of this is that
often if GDB times out trying to get a packet that feature will be
disabled.

This came up when I was looking at the test:

  gdb.threads/attach-many-short-lived-threads.exp

When running with the target-extended-remote board  this test
frequently times out on the attach.  Currently GDB just assumes that
the attach has succeeded and carries on, then it gets back the result
of the attach later, seemingly in reply to some future packet and gets
very confused.

Maybe it would be better to acknowledge that the attach failed??

TODO

 - think about this more, is this a good idea or not?