[WIP] gdb: better remote packet timeout support
Now instead of priming the buffer with null contents, place an error
string into the buffer.  If we get a timeout the packet will now be
considered an error.  This indicates that the packet is supported, but
the operation didn't succeed.

I've updated the test so that it spots the error emitted when the
timeout occurs.

Interestingly with this error in place, after the first iteration of
the test we now get an internal error (assertion) failure when we
attempt the second attach:

  thread.c:95: internal-error: thread_info* inferior_thread(): Assertion `current_thread_ != nullptr' failed.

So I really need to figure this out too....