gdb/riscv: store default target description in gdbarch
After the previous commit we are now able to store the default target
description within the gdbarch structure for riscv.

This is easy to do for the riscv architecture as we lookup an validate
the default description before looking up a previously created
gdbarch.  This is critical as now every gdbarch will have a target
description on it, either from the target, from the user, or a
default.

Many other architectures in gdb either create their gdbarch before
selecting a default target description, or they lookup a previously
created gdbarch before picking a default description.  None of these
reasons are totally blocking, however, with limited ability to test
all architectures I'm reluctant to make such a disruptive change.

So for now, everyone but riscv is left as they were, default target
descriptions are not stored within gdbarch.  This is harmless except
the commands 'maint print xml-tdesc' and 'maint print c-tdesc' will
not work when using a default description.

gdb/ChangeLog:

	* riscv-tdep.c (riscv_gdbarch_init): Ensure the default target
	description is placed into the gdbarch structure.